# apps/compass-ai/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Compass AI Backend API",
    "version": "2.1.0",
    "description": "Document analysis with grounded, citation-backed answers, scoped to a client portal. Tenancy is derived from a verified portalFeatureContextToken sent as x-portal-context: NO operation accepts a client, workspace or portal identifier from the caller, and one that carries one is rejected with TENANCY_PROBE. Every operation enforces the derived actor (employee or client) inside its handler. All operations are private: there is no cross-app consumer."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Every failure carries a stable machine code. For a portal failure the message is separately authored to be safe to show a client.",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "examples": [
                  "TENANCY_PROBE",
                  "PORTAL_NOT_BOUND",
                  "FORBIDDEN_CLIENT_ROLE",
                  "PORTAL_VERIFY_FAILED"
                ]
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
    },
    "parameters": {
      "PortalContext": {
        "name": "x-portal-context",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Platform-signed portalFeatureContextToken from the iframe URL. Verified through Gate on every request; a local decode is never trusted. It proves WHERE the request is from, never WHO — identity comes from the caller session."
      },
      "DocumentId": {
        "name": "documentId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "responses": {
      "TenancyProbe": {
        "description": "The request carried a client/workspace/portal identifier. Refused rather than ignored.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "PortalRejected": {
        "description": "Portal context missing, unverifiable, unbound, paused, or the caller is not a member.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ForbiddenClient": {
        "description": "FORBIDDEN_CLIENT_ROLE — this action is employee-only.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not found, or it belongs to another client. Deliberately indistinguishable.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/documents/{documentId}": {
      "get": {
        "operationId": "getDocument",
        "summary": "Read one document",
        "description": "A document of another client is a 404, never a 403.",
        "x-fusebase-visibility": "private",
        "x-fusebase-execution-mode": "sync",
        "parameters": [
          {
            "$ref": "#/components/parameters/PortalContext"
          },
          {
            "$ref": "#/components/parameters/DocumentId"
          }
        ],
        "responses": {
          "200": {
            "description": "Document"
          },
          "400": {
            "$ref": "#/components/responses/TenancyProbe"
          },
          "401": {
            "$ref": "#/components/responses/PortalRejected"
          },
          "403": {
            "$ref": "#/components/responses/PortalRejected"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/PortalRejected"
          }
        }
      },
      "delete": {
        "operationId": "deleteDocument",
        "summary": "Delete a document and its index entries",
        "description": "Employee-only and irreversible: pages, paragraphs and passages cascade from the document row.",
        "x-fusebase-visibility": "private",
        "x-fusebase-execution-mode": "sync",
        "parameters": [
          {
            "$ref": "#/components/parameters/PortalContext"
          },
          {
            "$ref": "#/components/parameters/DocumentId"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "400": {
            "$ref": "#/components/responses/TenancyProbe"
          },
          "401": {
            "$ref": "#/components/responses/PortalRejected"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenClient"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/PortalRejected"
          }
        }
      }
    },
...
    "/documents/{documentId}/reindex": {
      "post": {
        "operationId": "reindexDocument",
        "summary": "Re-run the pipeline from stored bytes",
        "description": "Employee-only.",
        "x-fusebase-visibility": "private",
        "x-fusebase-execution-mode": "async",
        "parameters": [
          {
            "$ref": "#/components/parameters/PortalContext"
          },
          {
            "$ref": "#/components/parameters/DocumentId"
          }
        ],
        "responses": {
          "202": {
            "description": "Enqueued"
          },
          "400": {
            "$ref": "#/components/responses/TenancyProbe"
          },
          "401": {
            "$ref": "#/components/responses/PortalRejected"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenClient"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/PortalRejected"
          }
        }
      }
    },
    "/folders": {
      "get": {
        "operationId": "listFolders",
        "summary": "The client’s folder tree",
        "description": "Both actors. `editable` states whether the caller may change the tree, so the UI does not infer it from the actor.",
        "x-fusebase-visibility": "private",
        "x-fusebase-execution-mode": "sync",
        "parameters": [
          {
            "$ref": "#/components/parameters/PortalContext"
          }
        ],
        "responses": {
          "200": {
            "description": "Folders and an editable flag"
          },
          "400": {
            "$ref": "#/components/responses/TenancyProbe"
          },
          "401": {
            "$ref": "#/components/responses/PortalRejected"
          },
          "403": {
            "$ref": "#/components/responses/PortalRejected"
          },
          "409": {
            "$ref": "#/components/responses/PortalRejected"
          }
        }
      },
      "post": {
        "operationId": "createFolder",
        "summary": "Create a folder",
        "description": "Employee-only. Five levels deep at most.",
        "x-fusebase-visibility": "private",
        "x-fusebase-execution-mode": "sync",
        "parameters": [
          {
            "$ref": "#/components/parameters/PortalContext"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "$ref": "#/components/responses/TenancyProbe"
          },
          "401": {
            "$ref": "#/components/responses/PortalRejected"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenClient"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "A folder of that name already exists here"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "parentId": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "uuid"
                  }
                },
                "required": [
                  "name"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/folders/documents/{documentId}": {
      "put": {
        "operationId": "moveDocument",
        "summary": "File a document into a folder",
        "description": "Employee-only. `folderId: null` moves it to Unfiled. Sets `folder_pinned`, so a source’s folder column stops re-filing it. Never re-indexes — a document’s location has nothing to do with its content.",
        "x-fusebase-visibility": "private",
        "x-fusebase-execution-mode": "sync",
        "parameters": [
          {
            "$ref": "#/components/parameters/PortalContext"
          },
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Moved"
          },
          "400": {
            "$ref": "#/components/responses/TenancyProbe"
          },
          "401": {
            "$ref": "#/components/responses/PortalRejected"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenClient"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/PortalRejected"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "folderId": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "uuid"
                  }
                },
                "required": [
                  "folderId"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/folders/{folderId}": {
      "patch": {
        "operationId": "updateFolder",
        "summary": "Rename and/or move a folder",
        "description": "Employee-only. `parentId` is applied only when the key is present, so a rename does not silently move the folder to the root. A move that would create a cycle or exceed the depth limit is a 400 naming which.",
        "x-fusebase-visibility": "private",
        "x-fusebase-execution-mode": "sync",
        "parameters": [
          {
            "$ref": "#/components/parameters/PortalContext"
          },
          {
            "name": "folderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated"
          },
          "400": {
            "$ref": "#/components/responses/TenancyProbe"
          },
          "401": {
            "$ref": "#/components/responses/PortalRejected"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenClient"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Name taken among siblings"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "parentId": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "uuid"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteFolder",
        "summary": "Delete a folder",
        "description": "Employee-only. Documents and subfolders move up to the parent; no document is ever deleted.",
        "x-fusebase-visibility": "private",
        "x-fusebase-execution-mode": "sync",
        "parameters": [
          {
            "$ref": "#/components/parameters/PortalContext"
          },
          {
            "name": "folderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "400": {
            "$ref": "#/components/responses/TenancyProbe"
          },
          "401": {
            "$ref": "#/components/responses/PortalRejected"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenClient"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/PortalRejected"
          }
        }
      }
    },
...
  }
}